refactored the controllers

+++ controlers/comment.js
@@ -1,4 +1,10 @@
-const { createComment, deleteComment, getCommentsByPostId, updateComment } = require("../db/comment");
+const {
+  getCommentById,
+  createComment,
+  deleteComment,
+  getCommentsByPostId,
+  updateComment,
+} = require("../db/comment");
 
 const commentController = {
   create: async (req, res, next) => {
@@ -8,45 +14,56 @@ const commentController = {
       req.body.content
     );
     if (comment) {
-      res.json({
+      return res.json({
         ok: true,
         comment,
       });
-      return;
     }
-    res.json({
+    return res.json({
       ok: false,
     });
   },
   delete: async (req, res, next) => {
+    const author = (await getCommentById(req.body.commentId)).author;
+
+    if (author.toString() !== req.user._id.toString()) {
+      return res.json({
+        ok: false,
+      });
+    }
     const comment = await deleteComment(req.body.commentId);
     if (comment) {
-      res.json({
+      return res.json({
         ok: true,
       });
-      return;
     }
-    res.json({
-      ok: false,
-    });
   },
   update: async (req, res) => {
     const content = req.body.content;
     const commentId = req.body.commentId;
 
-    const comment = await updateComment(commentId, {content: content, updatedAt: Date.now()});
+    const author = (await getCommentById(req.body.commentId)).author;
 
-    if(!comment){
-      return res.json({ok: false, comment});
+    if (author.toString() !== req.user._id.toString()) {
+      return res.json({ ok: false, comment });
     }
-    return res.json({ok: true, comment});
+
+    const comment = await updateComment(commentId, {
+      content: content,
+      updatedAt: Date.now(),
+    });
+
+    if (!comment) {
+      return res.json({ ok: false, comment });
+    }
+    return res.json({ ok: true, comment });
   },
   getCommentsByPostId: async (req, res, next) => {
     const comments = await getCommentsByPostId(req.params.postId);
-    res.json({
-      results: comments
+    return res.json({
+      results: comments,
     });
-  }
+  },
 };
 
 module.exports = commentController;
